Refactored a bit useCreateDateMutation.
